+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [feat]: `.equals()` method;
 - [feat]: `.cross()` method;
+- [feat]: `.dot()` method;
 
 ## 0.5.0 (2022-11-27)